+++ src/bl64-gcp-env.bash
@@ -1,0 +1,14 @@
+#######################################
+# BashLib64 / Module / Globals / Manage local filesystem
+#
+# Version: 1.0.0
+#######################################
+
+export BL64_GCP_CMD_GCLOUD="${BL64_GCP_CMD_GCLOUD:-}"
+
+export BL64_GCP_CONFIGURATION_NAME='bl64_gcp_configuration_private'
+export BL64_GCP_CONFIGURATION_CREATED="$BL64_LIB_VAR_FALSE"
+
+export BL64_GCP_SET_FORMAT_YAML=''
+export BL64_GCP_SET_FORMAT_TEXT=''
+export BL64_GCP_SET_FORMAT_JSON=''