+++ src/bl64-os-init.bash
@@ -1,7 +1,7 @@
 #######################################
 # BashLib64 / Module / Setup / OS / Identify OS attributes and provide command aliases
 #
-# Version: 1.0.0
+# Version: 1.2.0
 #######################################
 
 #######################################
@@ -21,75 +21,43 @@ function bl64_os_set_command() {
   # shellcheck disable=SC2034
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-*)
-    BL64_OS_CMD_AWK='/usr/bin/awk'
     BL64_OS_CMD_CAT='/bin/cat'
     BL64_OS_CMD_DATE='/bin/date'
     BL64_OS_CMD_FALSE='/bin/false'
-    BL64_OS_CMD_GAWK='/usr/bin/gawk'
-    BL64_OS_CMD_GREP='/bin/grep'
     BL64_OS_CMD_HOSTNAME='/bin/hostname'
     BL64_OS_CMD_ID='/usr/bin/id'
-    BL64_ARC_CMD_TAR='/bin/tar'
     BL64_OS_CMD_TRUE='/bin/true'
     BL64_OS_CMD_UNAME='/bin/uname'
     ;;
   ${BL64_OS_FD}-* | ${BL64_OS_CNT}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_OL}-*)
-    BL64_OS_CMD_AWK='/usr/bin/awk'
     BL64_OS_CMD_CAT='/usr/bin/cat'
     BL64_OS_CMD_DATE=/usr'/bin/date'
     BL64_OS_CMD_FALSE='/usr/bin/false'
-    BL64_OS_CMD_GAWK='/usr/bin/gawk'
-    BL64_OS_CMD_GREP='/usr/bin/grep'
     BL64_OS_CMD_HOSTNAME='/usr/bin/hostname'
     BL64_OS_CMD_ID='/usr/bin/id'
-    BL64_ARC_CMD_TAR='/bin/tar'
     BL64_OS_CMD_TRUE='/usr/bin/true'
     BL64_OS_CMD_UNAME='/bin/uname'
     ;;
   ${BL64_OS_ALP}-*)
-    BL64_OS_CMD_AWK='/usr/bin/awk'
     BL64_OS_CMD_CAT='/bin/cat'
     BL64_OS_CMD_DATE='/bin/date'
     BL64_OS_CMD_FALSE='/bin/false'
-    BL64_OS_CMD_GAWK='/usr/bin/gawk'
-    BL64_OS_CMD_GREP='/bin/grep'
     BL64_OS_CMD_HOSTNAME='/bin/hostname'
     BL64_OS_CMD_ID='/usr/bin/id'
-    BL64_ARC_CMD_TAR='/bin/tar'
     BL64_OS_CMD_TRUE='/bin/true'
     BL64_OS_CMD_UNAME='/bin/uname'
     ;;
   ${BL64_OS_MCOS}-*)
-    BL64_OS_CMD_AWK='/usr/bin/awk'
     BL64_OS_CMD_CAT='/bin/cat'
     BL64_OS_CMD_DATE='/bin/date'
     BL64_OS_CMD_FALSE='/usr/bin/false'
-    BL64_OS_CMD_GAWK="$BL64_LIB_INCOMPATIBLE"
-    BL64_OS_CMD_GREP='/usr/bin/grep'
     BL64_OS_CMD_HOSTNAME='/bin/hostname'
     BL64_OS_CMD_ID='/usr/bin/id'
-    BL64_ARC_CMD_TAR='/usr/bin/tar'
     BL64_OS_CMD_TRUE='/usr/bin/true'
     BL64_OS_CMD_UNAME='/usr/bin/uname'
     ;;
   *) bl64_check_show_unsupported ;;
   esac
-}
-
-#######################################
-# Create command sets for common options
-#
-# Arguments:
-#   None
-# Outputs:
-#   STDOUT: None
-#   STDERR: None
-# Returns:
-#   0: always ok
-#######################################
-# Warning: bootstrap function: use pure bash, no return, no exit
-function bl64_os_set_options() {
-  :
 }
 
 #######################################
@@ -108,24 +76,16 @@ function bl64_os_set_options() {
 #######################################
 # Warning: bootstrap function: use pure bash, no return, no exit
 function bl64_os_set_alias() {
-  local cmd_mawk='/usr/bin/mawk'
 
   # shellcheck disable=SC2034
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-* | ${BL64_OS_FD}-* | ${BL64_OS_CNT}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_OL}-*)
-    if [[ -x "$cmd_mawk" ]]; then
-      BL64_OS_ALIAS_AWK="$cmd_mawk"
-    else
-      BL64_OS_ALIAS_AWK="$BL64_OS_CMD_GAWK --traditional"
-    fi
     BL64_OS_ALIAS_ID_USER="${BL64_OS_CMD_ID} -u -n"
     ;;
   ${BL64_OS_ALP}-*)
-    BL64_OS_ALIAS_AWK="$BL64_OS_CMD_GAWK --traditional"
     BL64_OS_ALIAS_ID_USER="${BL64_OS_CMD_ID} -u -n"
     ;;
   ${BL64_OS_MCOS}-*)
-    BL64_OS_ALIAS_AWK="$BL64_OS_CMD_AWK"
     BL64_OS_ALIAS_ID_USER="${BL64_OS_CMD_ID} -u -n"
     ;;
   *) bl64_check_show_unsupported ;;